Move getdns_context_config into the library
