+++ src/getdns/getdns_extra.h.in
@@ -827,6 +827,9 @@ getdns_return_t getdns_str2list(const char *str, getdns_list **list);
 getdns_return_t getdns_str2bindata(const char *str, getdns_bindata **bindata);
 getdns_return_t getdns_str2int(const char *str, uint32_t *value);
 
+getdns_return_t
+getdns_context_config(getdns_context *context, const getdns_dict *config_dict);
+
 #ifdef __cplusplus
 }
 #endif
